Move normalization functions from general utils into easy/; create easy/utils since they may be used by multiple files in easy/
