+++ contextualized/easy/wrappers/test_normalize.py
@@ -7,7 +7,7 @@ import os
 import sys
 import numpy as np
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+sys.path.insert(0, os.path.abspath("../../.."))
 
 from contextualized.easy import ContextualizedRegressor
